requirements.txt
mcp
openai-whisper
